+++ tests/epyccel/test_epyccel_classes.py
@@ -49,18 +49,6 @@ def test_class_function(modnew):
     c = modnew.C()
     assert c.get_3() == 3
 
-@pytest.mark.parametrize( 'language', (
-        pytest.param("fortran", marks = [
-            pytest.mark.skip(reason="Argument reference count not increased. See #1586"),
-            pytest.mark.fortran]
-        ),
-        pytest.param("c", marks = [
-            pytest.mark.skip(reason="Argument reference count not increased. See #1586"),
-            pytest.mark.c]
-        ),
-        pytest.param("python", marks = pytest.mark.python)
-    )
-)
 def test_classes_1(language):
     import classes.classes_1 as mod
     modnew = epyccel(mod, language = language)
@@ -102,7 +90,7 @@ def test_classes_1(language):
     assert np.allclose(x1, x2, rtol=RTOL, atol=ATOL)
 
     l_py = mod.Line(p1_py)
-    l_l  = mod.Line(p1_l)
+    l_l  = modnew.Line(p1_l)
 
     assert p1_py.get_X() == p1_l.get_X()
     assert l_py.get_x() == l_l.get_x()