+++ CHANGELOG.md
@@ -42,6 +42,8 @@ All notable changes to this project will be documented in this file.
 -   Ensure `pyccel-init` calls the related function.
 -   Stop unnecessarily importing deprecated NumPy classes `int`, `bool`, `float`, `complex` in Python translation.
 -   #1712 : Fix library path and OpenMP support for recent Apple chips by getting Homebrew directory with `brew --prefix`.
+-   #1687 : Pointers in tuples are deallocated.
+-   #1586 : Raise an error for targets of class instances which go out of scope too early.
 
 ### Changed